Bug fix for CLI filter functions in speckle.py and cli.py
Bug fixes for kuan, frost and lee-enhanced speckle filters.
Functions had inconsistent type declarations for n_looks (must be integer) and damping_factor (must be Number or float). That issue apparently confused the CLI calls.

Also updated the cli.py to be consistent with the speckle code.

+++ eis_toolkit/cli.py
@@ -695,7 +695,7 @@ def lee_enhanced_filter_cli(
     input_raster: Annotated[Path, INPUT_FILE_OPTION],
     output_raster: Annotated[Path, OUTPUT_FILE_OPTION],
     size: int = 3,
-    n_looks: float = 1.0,
+    n_looks: int = 1,
     damping_factor: float = 1.0,
 ):
     """Apply an enhanced Lee filter to the input raster."""
@@ -723,7 +723,7 @@ def gamma_filter_cli(
     input_raster: Annotated[Path, INPUT_FILE_OPTION],
     output_raster: Annotated[Path, OUTPUT_FILE_OPTION],
     size: int = 3,
-    n_looks: float = 1.0,
+    n_looks: int = 1,
 ):
     """Apply a Gamma filter to the input raster."""
     from eis_toolkit.raster_processing.filters.speckle import gamma_filter
@@ -781,7 +781,7 @@ def kuan_filter_cli(
     input_raster: Annotated[Path, INPUT_FILE_OPTION],
     output_raster: Annotated[Path, OUTPUT_FILE_OPTION],
     size: int = 3,
-    n_looks: float = 1.0,
+    n_looks: int = 1,
 ):
     """Apply a Kuan filter to the input raster."""
     from eis_toolkit.raster_processing.filters.speckle import kuan_filter
